Move the getMaxLengths calculation into the table
