+++ src/viewModelFactories/tableViewModelFactory.ts
@@ -4,7 +4,6 @@ import { TableViewModel } from "../viewModels/tableViewModel";
 import { RowViewModel } from "../viewModels/rowViewModel";
 import { RowViewModelFactory } from "./rowViewModelFactory";
 import { RowViewModelFactoryParam } from "./rowViewModelFactoryParam";
-import { CellLengthCalculator } from "./cellLengthCalculator";
 
 export class TableViewModelFactory {
 
@@ -23,7 +22,7 @@ export class TableViewModelFactory {
             7) use the rest of the rows to get the contents via rowViewModelFactory.buildRow
             8) return the view model
         */
-        const maxColLengths: number[] = CellLengthCalculator.getMaxLengths(tableWithoutSeparator);
+        const maxColLengths: number[] = tableWithoutSeparator.getMaxLengths();
         //TODO: consider passing the entire table to the rowVmb.BuildRow method with the row index. 
         // Then the RowViewModelFactoryParam class can be deleted.
         let result = new TableViewModel();
