+++ test/cellLengthCalculator.test.ts
@@ -1,0 +1,10 @@
+import * as assert from 'assert';
+import { assertExt } from "./assertExtensions";
+import { PadCalculator } from './../src/viewModelFactories/padCalculator';
+
+suite("CellLengthCalculator tests", () => {
+
+    test("TODO", () => {
+        assert.fail(0, 1, "Write unit tests!", "");
+    });
+});
