--- test/viewModelFactories/cellLengthCalculator.test.ts
@@ -1,11 +1,0 @@
-import * as assert from 'assert';
-import { assertExt } from "../assertExtensions";
-import { RowViewModelFactoryParam } from '../../src/viewModelFactories/rowViewModelFactoryParam';
-import { PadCalculator } from '../../src/viewModelFactories/padCalculator';
-
-suite("CellLengthCalculator tests", () => {
-
-    test("TODO", () => {
-        assert.fail(0, 1, "Write unit tests!", "");
-    });
-});
